moving items into internal if they don't need to be consumed by a cli…
