+++ pkg/edgexsdk/edgexsdk.go
@@ -28,13 +28,13 @@ import (
 	"syscall"
 
 	"github.com/edgexfoundry/app-functions-sdk-go/internal"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/common"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/common"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/common/runtime"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger/http"
+	"github.com/edgexfoundry/app-functions-sdk-go/internal/trigger/messagebus"
 	"github.com/edgexfoundry/app-functions-sdk-go/pkg/excontext"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/runtime"
 	"github.com/edgexfoundry/app-functions-sdk-go/pkg/transforms"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger/http"
-	"github.com/edgexfoundry/app-functions-sdk-go/pkg/trigger/messagebus"
 	logger "github.com/edgexfoundry/go-mod-core-contracts/clients/logging"
 	registry "github.com/edgexfoundry/go-mod-registry"
 	"github.com/edgexfoundry/go-mod-registry/pkg/factory"
